KDS not dependent on ERA. I don't think that its going to be easy/feasible to use TRefs to point to objects in the ERA/rootevts files. There are a few reasons for this - the data are stored in Lyon and accessing the ROOT files requires a physical transfer of the data files from Lyon in order to open them up locally (unless there's some way of opening another shell and running commands from the shell); the other reason being that in order to use the TRefs, all of the rootevts files must be opened and rewritten in order to pass the UniqueID identifier of every pulse to Kds. This might be an option in the future, depending upon how the data acquisition progresses.
An new module, kraw has been created. This module does now depend upon ERA and the idea is that this objects in this module will fetch raw pulse traces for us. As of right now, however, there's not quite enough information in the output DST files to know exactly where the raw pulse traces are. But hopefully I can convince Eric to include that information into the rootevts files. 

The KPSA module, stands for Pulse Shape Analysis and right now just a bare-bones set of classes have been added. It will be a goal to make this module ROOT-independent, if possible, in order to be able to break off this module later to be used outside of Kdata. 

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/EdwDataStructure/trunk@608 e4a34bab-3947-0410-866c-ab59ef520b34