removed the kpsa_Dict.C/h files from the repository. ?\016those aren't supposed to be there. i must have accidentally added them in.
Added the KHalfComplexToRealDFT class. This has been verified to work in conjunction with the KRealToHalfComplexDFT class -- a fourier transform followed by an inverse fourier transform resulted in the same initial waveform to 1 part in 10^-12. 

deleted some lines that were already commented out in KPulseAnalysisChain - no operational change to this class. 

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/EdwDataStructure/trunk@614 e4a34bab-3947-0410-866c-ab59ef520b34