Move fPulseType and fChannelNumber from KBoloPulseRecord to KHLABoloPulseRecord.
This will require reprocessing of the KData and I should bump to a new version number after this. The reason for the change is that we shouldn't spend time worrying about the type of a pulse in the raw level file. that should be done afterward. by moving these variables out of the KBoloPulseRecord, we remove these variables from the raw level files. Currently they are in the HLA level records, but it might be better to put them into the Middle Level, which has yet to be built. 

Also, moved pulse type variable to the high level class (KHLABoloPulseRecord) since we need to do some data processing in order to determine which channels are collectrodes and which are vetos and to handle special voltage configurations properly.

KEventFactory  - removed some commented lines
KHLAEvent.h - made some private methods public

Added requirements for the jansson library for the kdatabase module

removed some comment lines from the various special Makefiles for different platforms

Added many many new values to the Raw/Base-class data structure that are filled by header information found in the Samba raw files. 

The mergeKEdsTree.cxx file has be explicitly removed from the build in the kds module. This file depends upon too many deprecated functions and is only useful from Run12. We should probably clean up KData soon to remove these old types of files. 

verified to run on Mac OS X, but need to verify on other platforms.


git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@970 e4a34bab-3947-0410-866c-ab59ef520b34