added a TSelector class CoverageSelector, which determines the coverage, the actual percentage of true events in the confidence region of a KQContourPoint, by generating random true and experimental events and checking if the true events whose corresponding experimental values are in a small frame in the E_recoil-Q-frame lie in the confidence region. This is a test of goodness of the determined confidence region.
added method KErecoilQDensity::SingleEventTrueProbDensity(Double_t* x,Double_t* par)
which describes the probability density f(a|x) for a true event a under the assumption of a measured event x
The uncertainties on the heat and ion channels (detector resolution) are dependent from the energies, i. e. from 
the true event and so the uncertainties are dependent from the arguments.
This is not the case for 
KERecoilQDensity::SingleEventProbDensity(Double_t* x,Double_t* par) 
which describes the likelihood f(x|a) for a measured event x under the assumption of a true event a
since the true event a which the uncertainties depend from is fixed

made KNeganovLuke methods inline since they have few code and might be called multiple times

some minor changes 

excluded kpta from Modules.mk as it seems not to compile at the moment


git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1237 e4a34bab-3947-0410-866c-ab59ef520b34