added a new method
KQContourPointList::GetDistributionOfTrueValuesMergeProb(...)

to determine the distribution of true events in a selected area.
This is done by integrating the countour functions of the single events 
in the list over a selected area in order to obtain the probabilities p_i, that the
true event which led to the corresponding measured event, is in the selected
area. Double_t arrays of size 2 representing the single event distributions
are created, a_0 = (1-p_a)  a_1 = p_a
 Then 2-event distributions are calculated from pairwise convoluting the
1-event distributions: c_n = sum_i a_i b_{n-i}
Then 4-event distributions are calculated by pairwise convoluting the 2-event
distributions the same way and so on, until the full list size is reached 

In order to integrate the likelihood/detector resolution function
over arbitraryly shaped areas, a modified version, where the user can
give a TF2 indicator function (1 in the area , 0 otherwise) has been created

KErecoilQDensity::SingleEventProbDensityWithIndicator(Double_t* x,Double_t* par)

added some documentation



git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1295 e4a34bab-3947-0410-866c-ab59ef520b34