updates
KChamonixKAmpSite / KOptimalKamper now handle shifting the pulse template. The KAmpEvent sets the samba pulse analysis record later in the method (why does this fix a bug… weird. that's a bug in itself). The PulseAnalysisRecord fExtra array was reduced in size because nobody is really using all of these data and I'm trying to figure out why the Amp level KData files are bigger and slower than I want them to be. MultiPeakFinder was added but it currently does nothing! the PulseShifter code was just slightly cleaned up for readability and I added a check at the beginning of the RunProcess method to not loop through the data if the shift was set to zero (this makes it fast if you want to turn shifting off without within a kampsite/kamper without having to recompile - just set the shift = 0 in your script.)

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1564 e4a34bab-3947-0410-866c-ab59ef520b34