feldberg update
reduced some overhead for-loops which are now only done once. cleaned up debug code. added a new KNullProcessor object that does nothing to an input pulse. This object can be useful when a Kamper or KAmpSite requires you to set a KPtaProcessor, but you want to control the KAmper/KAmpSite to not actually do anything.

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1625 e4a34bab-3947-0410-866c-ab59ef520b34