Major Update to Build System
This update brings KData to build on the latest versions of ROOT without requiring that ROOT to be installed from source. This build system now uses the $ROOTSYS/etc/Makefile.arch to direct the build. This update has been tested on Mac OS X (10.8), and the linux variants found on the CC in Lyon and Kalinka at KIT.

The new minimum version of ROOT is 5.32/00. Older versions of ROOT do not provide the ROOTSYS/etc/Makefile.arch file.

Additionally, for Mac users: Consider upgrading to 5.34/00 if you update your machine to 10.8 Mountain Lion. The reason is that the developer tools for Mountain Lion no longer include X11 by default. However, starting with ROOT version 5.34/00, ROOT does not need X11 as it uses the Cocoa framework (which looks really nice! see half-way down this page http://root.cern.ch/root/html534/notes/release-notes.html). In order to use the Cocoa framework with ROOT on Mac, configure with the --enable-cocoa option.

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1647 e4a34bab-3947-0410-866c-ab59ef520b34