root i/o for kpta
root i/o macros ( ClassDef and ClassImp) have been removed (commented out, actually) for all classes in the libkpta. The reason is that we do not need ROOT I/O Streamers for these classes because they should not be used to write any information to disk. Secondly, by removing these, this library maintains its portability should these classes want to be used in a different project, even one that does not use ROOT!  In fact, the kpta library could (and probably should) be maintained as an external project from KData.. Although, the libkamping library heavily depends on libkpta, so we keep the code together.

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1648 e4a34bab-3947-0410-866c-ab59ef520b34