dictionary generation
Dictionary Generation for libkamping is now using the "standard" ROOT way, by using a LinkDef.h file.  Other modules in KData don't use the LinkDef.h file, but rather using a special usage of rootcint.  Using LinkDef.h is more flexible and allows for proper usage of of the map<string,KResult> type in the ROOT CINT and in Python

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1655 e4a34bab-3947-0410-866c-ab59ef520b34