adding documentation to repository
instead of dealing with upgrading the svn machine to the latest python (which wasn't working) in order to install the proper python tools for automatic documentation creation, and since the documentation isn't that large, i'm adding the html docs to the repository directly. thus, everybody that owns a local copy of kdata also automatically has the latest docs. the next step will be to add the building of the documentation into the makefile. but i won't do that unless we have a problem of the documentation somehow becoming significantly outdated.

in order to rebuild the documentation, you need to do the following
1. $> root scripts/MakeHtml.C

2. $> cd htmlpy
    $> make

step 2 requires that you have a recent version of sphinx (a python doc tool) installed. to install sphinx, do the usual
$> sudo easy_install -U Sphinx

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1695 e4a34bab-3947-0410-866c-ab59ef520b34