last update
ionization channel analysis seems to work at the point where I'm going to check this into the repository. especially since ben might be interested in the tool.
however, i think there might be something fundamentally wrong with the optimal filter. i'm not certain. i think we need to run some simulation tests to see how well it performs and if the estimators are biased.

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1696 e4a34bab-3947-0410-866c-ab59ef520b34